新增我的文章功能（需額外新增一欄位，必執行artisan migrate。
更改環境：研究室

+++ app/Http/Controllers/PostsController.php
@@ -108,7 +108,14 @@ class PostsController extends Controller
             return redirect()->route('posts.index')
                              ->with('warning', '找不到該文章');
         }
+
+        if (is_null($post)) {
+            return redirect()->route('posts.index')
+                                ->with('warning', '您沒有權限編輯這文章');
+        }   
+
         $data = compact('post');
+
         return view('posts.edit', $data);
     }
 
@@ -130,6 +137,17 @@ class PostsController extends Controller
         foreach($post->comments as $comment) {
             $comment->delete();
         }
+
+        if (is_null($post)) {
+            return redirect()->route('posts,index')
+                                ->with('warning', '找不到該文章');
+        }
+
+        if ($post && $post->user->id != \Auth::user()->id) {
+            return redirect()->route('posts.index')
+                                ->with('warning', '您沒有權限刪除這篇文章');
+        }
+
         $post->delete();
         return redirect()->route('posts.index')
                          ->with('success', '刪除文章及留言成功');
@@ -137,7 +155,7 @@ class PostsController extends Controller
 
     public function comment($id, CommentRequest $request)
     {
-        $post    = \App\Post::find($id);
+        $post = \App\Post::find($id);
         
         if (is_null($post)) {
             return redirect()->route('posts.index')
